mlx5: consolidate device collection
Move device collection directly into Mlx5Collector instead of using a
single-use mixin. Replace untyped device dictionaries with a DeviceRecord
class whose mlx5_core_dev is required.

Keep low-level mlx5 object helpers in collect_device and convert records to
report dictionaries only at the output boundary.

Orabug: 39751588
